Move code to exec/any_sender_of.hpp
